refactor: view 클래스 패키지 변경에 따른 import문 변경
