+++ slipp/src/main/java/slipp/controller/LoginController.java
@@ -1,6 +1,6 @@
 package slipp.controller;
 
-import nextstep.mvc.tobe.JspView;
+import nextstep.mvc.tobe.view.JspView;
 import nextstep.mvc.tobe.ModelAndView;
 import nextstep.web.annotation.Controller;
 import nextstep.web.annotation.RequestMapping;
